backend/requirements_supabase.txt
fastapi
uvicorn
pdfplumber
python-multipart
python-dotenv
supabase==2.10.0
